Rider
We have a road in the form of a ring, where the driver goes. The road length is 109 km. the Code shows which kilometer of road the driver is on. If the speed is set to a negative value, then the driver is traveling in the opposite direction. 

These are the tasks we usually have)